phase-3/backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlmodel==0.0.14
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
pyjwt>=2.10.1
bcrypt==4.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Configuration
python-dotenv==1.0.0

# Validation
pydantic[email]>=2.10.0
pydantic-settings>=2.1.0

# CORS
python-multipart>=0.0.9

# AI Agent Framework (Phase 3)
openai-agents[litellm]>=0.1.0

# MCP Integration (Phase 3)
mcp>=1.0.0

# LLM Integration (Phase 3)
groq>=0.4.0
litellm>=1.0.0

# Token Counting (Phase 3)
tiktoken>=0.5.0

# HTTP Client (Phase 3)
httpx>=0.25.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
